+++ aldrin/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Multiple service ids can now be queried on `Discoverer` and related types.
 - Add `UnknownCall::take_args()`, `into_promise()` and `into_args_and_promise()`.
 - Add `low_level::Call::promise()` and `promise_mut()`.
+- Add `low_level::Call::set_as()`, `set()`, `ok_as()`, `ok()`, `done()`, `err_as()` and `err()`.
 
 ### Changed